src/requirements.txt
Flask==3.0.1
flask-cors==4.0.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
requests==2.31.0
python-dotenv==1.0.1
joblib==1.3.2
PyJWT==2.8.0
